Add vinebot application locks for multiple leaves
When multiple components are connected to ejabberd at the same subdomain, ejabberd round-robbins traffic between them. I want to take advantage of this load-balancing behavior to make my leaf components more scalable and robust (so I can restart one without the whole application going offline).

In order for this to work, however, I have to make sure that two leaves aren't simultaneously handling stanzas for the same vinebot. If this were to happen, then weird race conditions could occur, because I have two databases (mysql and mnesia). What happens, for example, if someone is being added to a two-person conversation, but before that finishes, someone else leaes? What is the end result for that contact on each user's roster?

To avoid these issues, I want to use mysql application locks so that only one leaf is able to interact with both the mysql and mnesia data for each vinebot at a time. In the constructor for the vinebot, which is called at the beginning of each handle_ funciton, a lock is acquired, and then it is released at the end. Things are a little more complicated for adding/removing participants, and some of the admin /commands, since multiple vinebots are being written to for the same stanza, but it's mostly the same.

NOTE: This code is broken, because when a single connection that has already acquired a lock with get_lock calls get_lock again to acquire a second lock, the first lock is released. This is a problem, and I don't know how to solve it, but this commit is otherwise relatively self-contained and clean.

+++ edge.py
@@ -4,6 +4,7 @@ import sys
 import logging
 from constants import g
 import user as u
+import vinebot as v
 
 if sys.version_info < (3, 0):
     reload(sys)
@@ -22,6 +23,8 @@ class AbstractEdge(object):
         self.id = None
     
     def change_vinebot(self, vinebot):
+        if not vinebot.can_write:
+            raise v.VinebotPermissionsException
         g.db.execute("""UPDATE edges
                         SET vinebot_id = %(vinebot_id)s
                         WHERE id = %(id)s
@@ -31,7 +34,9 @@ class AbstractEdge(object):
                      })
         self.vinebot_id = vinebot.id
     
-    def delete(self):
+    def delete(self, vinebot):  # passing in this vinebot doesn't actually protect us from anything, but forces us to be explicit
+        if not vinebot.can_write:
+            raise v.VinebotPermissionsException
         g.db.execute("""DELETE FROM edges
                             WHERE id = %(id)s
                          """, {           
@@ -46,18 +51,20 @@ class AbstractEdge(object):
     
 
 class InsertedEdge(AbstractEdge):
-    def __init__(self, f_user, t_user, vinebot_id):
+    def __init__(self, f_user, t_user, vinebot):        
+        if not vinebot.can_write:
+            raise v.VinebotPermissionsException
         super(InsertedEdge, self).__init__()
         dbid = g.db.execute("""INSERT INTO edges (from_id, to_id, vinebot_id)
                                    VALUES (%(f_id)s, %(t_id)s, (%(vinebot_id)s))
                                 """, {           
                                    't_id': t_user.id, 
                                    'f_id': f_user.id,
-                                   'vinebot_id': vinebot_id
+                                   'vinebot_id': vinebot.id
                                 })
         self.f_user = f_user
         self.t_user = t_user
-        self.vinebot_id = vinebot_id
+        self.vinebot_id = vinebot.id
         self.id = dbid
     
 
@@ -159,4 +166,4 @@ class FetchedEdge(AbstractEdge):
                                       dbid=result[2]
                                      ) for result in results]
         return frozenset(to_edges + from_edges)
-   
+
